usdriver.py -- component name changes, added initialise overide option to menu
The settings menu now contains an item to allow one to assume that the servers have been initialised. This is to cope with the case that one has got out of usdriver and restarted it while the servers continue to run. Probably mainly for the experts hence it is only enabled in expert mode. Components renamed to reflect the purpose more accurately.