uspec.py -- fixes for disabling of target, PI buttons
According to the data type, some buttons are frozen. e.g. if you select bias, it will make the name "Bias" and set the PI and Proposal ID to Calib. This was not quite working. I also found that the target lines were commented out in the "check" routine and I uncommented them, but there may have been a reason.